test with firefox, fix number scroll test
on firefox, the window width is much larger than on chrome. window dimensions also cannot be easily set from karma. because of the larger window, the margin between the document width and the viewport width is very small, leading to the horizontal scroll wiht number target test to fail. the issue was due to the "stretching" element using vh units rather than vw units (which on a typical landscape orientation does not result in a wide width). switching to vw fixed the issue without needing to set window size or dynamically set scroll target based on window size.